refactor: 해당 type model 패키지에서 type으로 옮김
