+++ src/main/java/com/project/trainingdiary/model/type/NotificationType.java
@@ -1,4 +1,4 @@
-package com.project.trainingdiary.model;
+package com.project.trainingdiary.model.type;
 
 public enum NotificationType {
