+++ src/main/java/com/project/trainingdiary/model/type/ScheduleStatusType.java
@@ -1,0 +1,8 @@
+package com.project.trainingdiary.model.type;
+
+public enum ScheduleStatusType {
+
+  OPEN,
+  RESERVE_APPLIED,
+  RESERVED
+}
