+++ src/main/java/com/project/trainingdiary/model/type/UserRoleType.java
@@ -1,4 +1,4 @@
-package com.project.trainingdiary.model;
+package com.project.trainingdiary.model.type;
 
 public enum UserRoleType {
